app/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-cors-headers==3.10.0
django-crispy-forms
djangorestframework==3.12.4
django-filter==2.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
pycodestyle==2.6.0
psycopg2==2.8.4
psycopg2-binary==2.9.1
drf-nested-routers==0.93.3
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
django-debug-toolbar